helper functions, move pin map LUT into .c
